requirement.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pypdf
python-dotenv
numpy
protobuf==3.20.3

# LangGraph + LangChain 1.x
langgraph>=1.0.5
langchain>=1.2.3,<2.0
langchain-core>=1.2.7,<2.0
langchain-community>=0.4.1
langchain-text-splitters>=1.1.0,<2.0
langsmith>=0.6.2


# OpenAI / Grok support

langchain-groq
